+++ src/dms/backend/EdFi.DataManagementService.Backend.Postgresql/Deploy/Scripts/0004_Create_EducationOrganizationHierarchy_Table.sql
@@ -7,7 +7,7 @@ CREATE TABLE dms.EducationOrganizationHierarchy(
     Id BIGINT UNIQUE GENERATED ALWAYS AS IDENTITY(START WITH 1 INCREMENT BY 1),
     ProjectName VARCHAR(256) NOT NULL,
     ResourceName VARCHAR(256) NOT NULL,
-    EducationOrganizationId INT NOT NULL,
+    EducationOrganizationId BIGINT NOT NULL,
     ParentId BIGINT REFERENCES dms.EducationOrganizationHierarchy(Id) ON DELETE CASCADE,
     UNIQUE (ProjectName, ResourceName, EducationOrganizationId, ParentId)
 );
